requirements.txt
# Веб-фреймворк
Flask==3.0.0
Flask-SQLAlchemy==3.1.1          # ORM для работы с базой данных
Flask-Login==0.6.3               # Управление сессиями пользователей
Flask-WTF==1.2.1                 # Формы с защитой CSRF
Werkzeug==3.0.1                  # WSGI утилиты (хеширование паролей)

# Обработка изображений и OCR
Pillow==10.1.0                   # Библиотека для работы с изображениями
opencv-python==4.8.1.78          # Компьютерное зрение для обработки изображений
PyMuPDF==1.23.8                  # Конвертация PDF в изображения
numpy==1.26.2                    # Работа с массивами (для OpenCV)

# Работа с документами
PyPDF2==3.0.1                    # Чтение и запись PDF
python-docx==1.1.0               # Создание и редактирование DOCX
reportlab==4.0.7                 # Генерация PDF документов
openpyxl==3.1.2                  # Работа с Excel (опционально)

# Утилиты
python-dotenv==1.0.0             # Загрузка переменных окружения
python-magic==0.4.27             # Определение типов файлов
validators==0.22.0               # Валидация данных

PyMuPDF==1.23.8
easyocr==1.7.0
torch==2.1.0  # для EasyOCR
torchvision==0.16.0
